+++ spec/lib/ordnance_survey/configuration/client_spec.rb
@@ -1,0 +1,19 @@
+require "rails_helper"
+
+module OrdnanceSurvey
+  RSpec.describe Configuration::Client do
+    subject(:client) { described_class.new }
+
+    describe "#base_url" do
+      it "returns nil" do
+        expect(client.base_url).to eq(nil)
+      end
+    end
+
+    describe "#params" do
+      it "returns empty hash" do
+        expect(client.params).to eq({})
+      end
+    end
+  end
+end